refactor: 엔티티 연관관계 수정
- 패키지 하위에 단일 클래스파일인 경우 패키지 제거
- Account, AccountTag 양방향 연관관계 설정
- Pet, PetTag 양방향 연관관계 설정
- AccountImage, Account 연관관계 제거
- PetImage, Pet 연관관계 제거
