+++ src/main/java/com/yapp/pet/domain/pet_image/PetImage.java
@@ -1,15 +1,12 @@
-package com.yapp.pet.domain.pet_image.entity;
+package com.yapp.pet.domain.pet_image;
 
 import com.yapp.pet.domain.common.BaseEntity;
-import com.yapp.pet.domain.pet.entity.Pet;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-
-import static javax.persistence.FetchType.LAZY;
 
 @Entity
 @Getter
@@ -30,15 +27,10 @@ public class PetImage extends BaseEntity {
     @Column(nullable = false)
     private String path;
 
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "pet_id")
-    private Pet pet;
-
     @Builder
-    public PetImage(String originName, String name, String path, Pet pet) {
+    public PetImage(String originName, String name, String path) {
         this.originName = originName;
         this.name = name;
         this.path = path;
-        this.pet = pet;
     }
 }
