+++ src/main/java/com/yapp/pet/domain/pet_tag/PetTag.java
@@ -1,4 +1,4 @@
-package com.yapp.pet.domain.pet_tag.entity;
+package com.yapp.pet.domain.pet_tag;
 
 import com.yapp.pet.domain.common.BaseEntity;
 import com.yapp.pet.domain.pet.entity.Pet;
@@ -13,7 +13,6 @@ import static javax.persistence.FetchType.LAZY;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "pet_prefer_tag")
 public class PetTag extends BaseEntity {
 
     @Id
@@ -21,19 +20,20 @@ public class PetTag extends BaseEntity {
     @Column(name = "pet_tag_id")
     private Long id;
 
-    @Column(nullable = false, length = 10)
-    private String name;
-
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "pet_id")
     private Pet pet;
 
-    private PetTag(String name, Pet pet) {
-        this.name = name;
+    @Column(nullable = false, length = 10)
+    private String name;
+
+    private PetTag(Pet pet, String name) {
         this.pet = pet;
+        this.name = name;
     }
 
-    public static PetTag of(String name, Pet pet) {
-        return new PetTag(name, pet);
+    public static PetTag of(Pet pet, String name) {
+        return new PetTag(pet, name);
     }
+
 }
